chore: abstract out questionpanel and roomoptions
