+++ frontend/components/room/QuestionPanel/index.ts
@@ -1,0 +1,2 @@
+import QuestionPanel from './QuestionPanel';
+export default QuestionPanel;
