+++ frontend/components/room/RoomOptions/index.ts
@@ -1,0 +1,2 @@
+import RoomOptions from './RoomOptions';
+export default RoomOptions;
